+++ vantage-sql/src/mysql/table_source.rs
@@ -485,6 +485,10 @@ impl TableSource for MysqlDB {
         tgt_col.in_(fk_values.expr())
     }
 
+    fn supports_traversal(&self) -> bool {
+        true
+    }
+
     fn related_correlated_condition(
         &self,
         target_table: &str,