+++ vantage-sql/tests/sqlite.rs
@@ -22,6 +22,8 @@ mod sqlite {
     mod expressions;
     #[path = "2_identifier.rs"]
     mod identifier;
+    #[path = "8_implicit_references.rs"]
+    mod implicit_references;
     #[path = "2_insert.rs"]
     mod insert;
     #[path = "5_invariants.rs"]